refactor(LobbyScreen):exported RolesList as separate module
